Remove unnecessary models, only use Collection
Without this commit, each new collection type would need a new serializer, a new model, a new view on the backend, and new frontend templates. This becomes unscalable especially as we open up collections to other people, and is already unwieldy for developers juggling only a few collection types.

This commit factors out CollectionBase and Meeting as models, with everything from both being placed in Collection. In the future, most of the metadata in collections will be decided based on customizable JSON schema that can be added to collections and their items. This commit lays the foundation for the implementation of modular metadata.