Remove group serializers, views, and model
This commit fully removes Groups from the API. This has been done for several reasons:
1) They are no longer being used on the frontend in any capacity
2) Cameron (@cameronblandford) has deemed them to no longer be a useful abstraction, with agreement from teammates
3) The extra code bloats and needlessly complicates things. Not having that extra unnused code makes the codebase more streamlined, purposeful, and clean.

While this decision may be reverted in the future, the groups code was stale anyway, and would need significant revamping to be useful again. Should someone choose to re-introduce groups, this will give them a clean slate with which to do so.